add watch mode
