+++ packages/reciple/package.json
@@ -44,6 +44,7 @@
     "-": "^0.0.1",
     "@reciple/update-checker": "^9.0.0",
     "@reciple/utils": "^8.2.1",
+    "chokidar": "^3.6.0",
     "commander": "^12.0.0",
     "dotenv": "^16.4.5",
     "fallout-utility": "^2.8.0",
