+++ packages/reciple/src/bin.mts
@@ -1,18 +1,19 @@
 #!/usr/bin/env node
 
-import { ContextMenuCommandBuilder, Logger, MessageCommandBuilder, SlashCommandBuilder, buildVersion } from '@reciple/core';
-import { createLogger, addEventListenersToClient } from './utils/logger.js';
-import { setTimeout as setTimeoutAsync } from 'node:timers/promises';
-import { command, cli, cliVersion, checkForUpdates } from './utils/cli.js';
-import { ProcessInformation, RecipleClient, findModules } from './index.js';
+import { Logger, buildVersion } from '@reciple/core';
+import { createLogger } from './utils/logger.js';
+import { command, cli, cliVersion } from './utils/cli.js';
+import { ProcessInformation, RecipleClient } from './index.js';
 import { existsAsync, resolveEnvProtocol } from '@reciple/utils';
 import { ConfigReader } from './classes/Config.js';
 import { config as loadEnv } from 'dotenv';
 import { mkdir } from 'node:fs/promises';
-import { kleur } from 'fallout-utility';
 import { parentPort, threadId } from 'node:worker_threads';
 import path from 'node:path';
 import semver from 'semver';
+import { watch } from 'chokidar';
+import { createClient } from './utils/modules.js';
+import { spawn } from 'node:child_process';
 
 command.parse();
 
@@ -49,9 +50,7 @@ if (cli.shardmode) config.applicationCommandRegister = { ...config.applicationCo
 if (cli.options.token) config.token = resolveEnvProtocol(cli.options.token) ?? config.token;
 
 const processErrorHandler = (err: any) => logger?.error(err);
-
-process.once('uncaughtException', processErrorHandler);
-process.once('unhandledRejection', processErrorHandler);
+const watcher = config.watch?.enabled ? watch(config.watch?.include ?? '.') : null;
 
 process.on('warning', warn => logger?.warn(warn));
 
@@ -69,92 +68,41 @@ if (config.version && !semver.satisfies(cliVersion, config.version)) {
 
 logger?.info(`Starting Reciple client v${buildVersion} - ${new Date()}`);
 
-const client = new RecipleClient(config);
+let client = new RecipleClient(config);
 
-client.setLogger(logger);
+await initializeClient();
 
-addEventListenersToClient(client);
+watcher?.on('all', async event => {
+    if (config.watch?.reloadTriggerEvent && config.watch?.reloadTriggerEvent !== 'all' && config.watch?.reloadTriggerEvent !== event) return;
 
-const moduleFilesFilter = (file: string) => file.endsWith('.js') || file.endsWith('.cjs') || file.endsWith('.mjs');
+    if (config.watch?.preLoadScript) {
+        const child = spawn('npm', ['run', config.watch.preLoadScript], {
+            cwd: cli.cwd,
+            env: process.env,
+            stdio: ['ignore', 'inherit', 'inherit']
+        });
 
-const modules = await client.modules.resolveModuleFiles({
-    files: await findModules(config.modules, (f) => moduleFilesFilter(f)),
-    disableVersionCheck: config.modules?.disableModuleVersionCheck
-});
+        const code = await new Promise(res => child.on('exit', code => res(code)));
+        if (code) {
+            logger?.error(`Watch pre-load script exited with code (${code})`);
+            return;
+        }
+    }
 
-const startedModules = await client.modules.startModules();
-const failedToStartModules = modules.length - startedModules.length;
+    await initializeClient();
+})
 
-if (failedToStartModules > 0) logger?.error(`Failed to start (${failedToStartModules}) modules.`);
+async function initializeClient() {
+    process.once('uncaughtException', processErrorHandler);
+    process.once('unhandledRejection', processErrorHandler);
 
-client.once('ready', async () => {
-    if (!client.isReady()) return;
+    if (client) {
+        await client.destroy(true);
+        console.clear();
+    }
 
-    logger?.debug(`Client is ready!`);
+    client = await createClient(config, logger);
 
     process.removeListener('uncaughtException', processErrorHandler);
     process.removeListener('unhandledRejection', processErrorHandler);
-
-    const loadedModules = await client.modules.loadModules({ cacheCommands: true });
-    const failedToLoadModules = startedModules.length - loadedModules.length;
-
-    if (failedToLoadModules > 0) logger?.debug(`Failed to load (${failedToLoadModules}) modules.`);
-
-    let stopping = false;
-
-    const unloadModulesAndStopProcess = async (signal: NodeJS.Signals) => {
-        if (stopping) return;
-
-        logger?.debug(`Received exit signal: ${signal}`);
-
-        stopping = true;
-
-        await client.destroy(true);
-
-        const signalString = signal === 'SIGINT' ? 'keyboard interrupt' : signal === 'SIGTERM' ? 'terminate' : String(signal);
-
-        logger?.warn(`Process exited: ${kleur.yellow(signalString)}`);
-        logger?.closeWriteStream();
-
-        await setTimeoutAsync(10);
-        process.exit(0);
-    };
-
-    process.stdin.resume();
-
-    process.once('SIGHUP', unloadModulesAndStopProcess);
-    process.once('SIGINT', unloadModulesAndStopProcess);
-    process.once('SIGQUIT', unloadModulesAndStopProcess);
-    process.once('SIGABRT',unloadModulesAndStopProcess);
-    process.once('SIGALRM', unloadModulesAndStopProcess);
-    process.once('SIGTERM', unloadModulesAndStopProcess);
-    process.once('SIGBREAK', unloadModulesAndStopProcess);
-    process.once('SIGUSR2', unloadModulesAndStopProcess);
-
-    client.on('interactionCreate', interaction => {
-        if (interaction.isContextMenuCommand()) {
-            ContextMenuCommandBuilder.execute({ client, interaction });
-        } else if (interaction.isChatInputCommand()) {
-            SlashCommandBuilder.execute({ client, interaction });
-        }
-    });
-
-    client.on('messageCreate', message => {
-        MessageCommandBuilder.execute({ client, message });
-    });
-
-    if (config.applicationCommandRegister?.enabled !== false) await client.commands.registerApplicationCommands();
-
-    logger?.warn(`Logged in as ${kleur.bold().cyan(client.user.tag)} ${kleur.magenta('(' + client.user.id + ')')}`);
-
-    logger?.log(`Loaded ${client.commands.contextMenuCommands.size} context menu command(s)`);
-    logger?.log(`Loaded ${client.commands.messageCommands.size} message command(s)`);
-    logger?.log(`Loaded ${client.commands.slashCommands.size} slash command(s)`);
-    logger?.log(`Loaded ${client.commands.preconditions.size} precondition(s)`);
-
-    if (config.checkForUpdates) await checkForUpdates(logger ?? undefined);
-});
-
-logger?.debug(`Logging in...`);
-
-await client.login().then(() => logger?.debug(`Login successful`));
+}
